picsaver/requirements.txt
Flask==0.10.1
Flask-Pagination==0.0.1
httplib2==0.9.2
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
python-instagram==1.3.2
simplejson==3.8.1
six==1.10.0
Werkzeug==0.11.2
wheel==0.24.0
